Title shortening added.
There are books which have utf-8 representation of title longer than
240 bytes. Thus filename made from book title may exceed filesystem
limit,

Prevoisly booksort just throwed an exception in this case and skipped
file letting user to deal with problem manually.

Now it tries to find delimiter between title and subtitle, and use
just title as filename, leaving metainformation in the file intact.